:truck: Rename SheetView to SheetPage
